Moved utility next to HtmlCaptureExtension
