+++ cubano-concordion/src/main/java/org/concordion/cubano/driver/concordion/ExceptionHtmlCaptureExtension.java
@@ -4,9 +4,9 @@ import org.concordion.api.extension.ConcordionExtender;
 import org.concordion.api.extension.ConcordionExtension;
 import org.concordion.api.listener.ThrowableCaughtEvent;
 import org.concordion.api.listener.ThrowableCaughtListener;
+import org.concordion.cubano.driver.concordion.pagegrabber.GrabWebPage;
 import org.concordion.cubano.driver.http.HttpDownloader;
 import org.concordion.cubano.driver.web.Browser;
-import org.concordion.cubano.driver.web.pagegrabber.GrabWebPage;
 import org.concordion.ext.StoryboardExtension;
 import org.concordion.ext.storyboard.CardResult;
 import org.openqa.selenium.NoSuchElementException;
