+++ cubano-concordion/src/main/java/org/concordion/cubano/driver/concordion/pagegrabber/GrabUtility.java
@@ -1,4 +1,4 @@
-package org.concordion.cubano.driver.web.pagegrabber;
+package org.concordion.cubano.driver.concordion.pagegrabber;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
