+++ cubano-concordion/src/main/java/org/concordion/cubano/driver/concordion/pagegrabber/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * PageGrabber utility.
+ */
+package org.concordion.cubano.driver.concordion.pagegrabber;
