--- cubano-webdriver/src/main/java/org/concordion/cubano/driver/web/pagegrabber/package-info.java
@@ -1,4 +1,0 @@
-/**
- * PageGrabber utility.
- */
-package org.concordion.cubano.driver.web.pagegrabber;
